Instance: creata a new delegated systemd scope
This forces systemd to enable all cgroup controllers it supports.
That is important because all controllers used by beng-proxy, Workshop
and OpenSSH must be cleaned up, and not only the ones which are
enabled by default.

Delegated scopes are supported since systemd 117; just to be sure,
raise the systemd dependency to 229, which is the systemd version in
Ubuntu 16.04LTS; Debian Stretch has version 230.